Add files via upload
Analyzing both the date and time of crime occurrences, we have identified two key charts that offer crucial insights. The good news is that 2024 experiences a huge decline in crime rates. The time chart may seem "odd," but it is the most significant. It shows that 12:00 PM has the highest crime rate of the day, which may seem unrealistic at first glance. Further inspection is needed to confirm whether this is accurate or if it indicates a data entry error.